implement authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -53,16 +53,23 @@ class QuestionsController extends Controller
 
     public function edit(Question $question)
     {
-        if(\Gate::allows('update-question', $question)) 
+        if(\Gate::denies('update-question', $question)) 
         {
-            return view("questions.edit", compact('question'));
+            abort(403, "Access Denied! Go back man!");
         }
-        abort(403, "Access Denied");
+
+        return view("questions.edit", compact('question'));
+        
     }
 
 
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)) 
+        {
+            abort(403, "Access Denied! Go back man!");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -71,6 +78,11 @@ class QuestionsController extends Controller
 
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)) 
+        {
+            abort(403, "Access Denied! Go back man!");
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
